backend/requirements.txt
altgraph==0.17.5
asgiref==3.11.0
attrs==25.4.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pefile==2024.8.26
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.10.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
waitress==3.0.2
whitenoise==6.11.0
